add strings
