+++ checkout/src/main/java/com/mercadopago/sdk/android/checkout/domain/extensions/FieldStateExtensions.kt
@@ -1,0 +1,5 @@
+package com.mercadopago.sdk.android.checkout.domain.extensions
+
+import com.mercadopago.sdk.android.checkout.presentation.state.CardNumberState
+
+internal fun CardNumberState.isComplete() = length == maxLength
